additional tuning parameters added to AcceleratedModule class
Former-commit-id: 326df209bfefd4728c591f9f99cd25237ed61397
Former-commit-id: e12fdeef5e0060fc08656adfa2f215cf5b26fdc5

+++ modules/AccelerationModule.py
@@ -78,7 +78,7 @@ class AccelerationModule(object):
         min_eig = RNA.min_eignevalRR(x_hist_np)
         return min_eig
 
-    def accelerate(self, model, step_size: int = 1):
+    def accelerate(self, model, step_size: float = 1):
 
         if len(self.x_hist) < 3:  # Cannot accelerate when number of points is below 3
             self.load_param_in_model(np.array(self.x_hist[-1]), model)
